Support db.rollback in every commit

+++ appcomposer/translator/ops.py
@@ -27,6 +27,10 @@ def get_golab_default_user():
             db.session.commit()
         except IntegrityError:
             default_user = db.session.query(GoLabOAuthUser).filter_by(email = default_email).first()
+            db.session.rollback()
+        except:
+            db.session.rollback()
+            raise
     return default_user
 
 def _get_or_create_app(app_url, translation_url):
@@ -89,8 +93,7 @@ def add_full_translation_to_app(user, app_url, translation_url, language, target
         category = original_messages.get(key, {}).get('category')
         if category is not None and existing_active_translation.category != category:
             existing_active_translation.category = category
-
-
+    
     if translated_messages is not None:
         # Delete active translations that are going to be replaced
         # Store which were the parents of those translations and
@@ -101,7 +104,7 @@ def add_full_translation_to_app(user, app_url, translation_url, language, target
         for existing_active_translation in db.session.query(ActiveTranslationMessage).filter_by(bundle = db_translation_bundle).all():
             key = existing_active_translation.key
             if key in translated_messages:
-                if existing_active_translation.value != translated_messages[key] or (not from_developer and existing_active_translation.taken_from_default):
+                if (translated_messages[key] and existing_active_translation.value != translated_messages[key]) or (not from_developer and existing_active_translation.taken_from_default):
                     parent_translation_ids[key] = existing_active_translation.history.id
                     db.session.delete(existing_active_translation)
                 else:
@@ -110,6 +113,9 @@ def add_full_translation_to_app(user, app_url, translation_url, language, target
         # For each translation message
         now = datetime.datetime.utcnow()
         for key, value in translated_messages.iteritems():
+            if value is None:
+                value = ""
+
             if key not in unchanged and key in original_messages:
                 # Create a new history message
                 parent_translation_id = parent_translation_ids.get(key, None)
@@ -157,11 +163,14 @@ def add_full_translation_to_app(user, app_url, translation_url, language, target
         except IntegrityError:
             # Somebody else concurrently run this
             db.session.rollback() 
+        except:
+            db.session.rollback()
+            raise
 
     now = datetime.datetime.utcnow()
     existing_keys = [ key for key, in db.session.query(ActiveTranslationMessage.key).filter_by(bundle = db_translation_bundle).all() ]
     for key, original_message_pack in original_messages.iteritems():
-        value = original_message_pack['text']
+        value = original_message_pack['text'] or ''
         position = original_message_pack['position']
         category = original_message_pack['category']
         if key not in existing_keys:
@@ -185,6 +194,9 @@ def add_full_translation_to_app(user, app_url, translation_url, language, target
     except IntegrityError:
         # Somebody else did this
         db.session.rollback()
+    except:
+        db.session.rollback()
+        raise
     else:
         from appcomposer.translator.tasks import push_task
         push_task.delay(translation_url, language, target)
@@ -196,6 +208,9 @@ def register_app_url(app_url, translation_url):
     except IntegrityError:
         # Somebody else did this process
         db.session.rollback()
+    except:
+        db.session.rollback()
+        raise
     else:
         # Delay the synchronization process
         from appcomposer.translator.tasks import synchronize_apps_no_cache_wrapper
@@ -385,7 +400,11 @@ def _deep_copy_bundle(src_bundle, dst_bundle):
     for msg in src_bundle.all_messages:
         t_history = TranslationMessageHistory(dst_bundle, msg.key, msg.value, msg.user, msg.datetime, src_message_ids.get(msg.parent_translation_id), msg.taken_from_default)
         db.session.add(t_history)
-        db.session.commit()
+        try:
+            db.session.commit()
+        except:
+            db.session.rollback()
+            raise
         db.session.refresh(t_history)
         src_message_ids[msg.id] = t_history.id
         historic[msg.id] = t_history
@@ -396,7 +415,11 @@ def _deep_copy_bundle(src_bundle, dst_bundle):
         active_t = ActiveTranslationMessage(dst_bundle, msg.key, msg.value, history, now, msg.taken_from_default, msg.position, msg.category, msg.from_developer)
         db.session.add(active_t)
 
-    db.session.commit()
+    try:
+        db.session.commit()
+    except:
+        db.session.rollback()
+        raise
 
 def _merge_bundle(src_bundle, dst_bundle):
     """Copy all the messages. The destination bundle already existed, so we can only copy those
@@ -409,7 +432,11 @@ def _merge_bundle(src_bundle, dst_bundle):
             db.session.add(t_history)
             active_t = ActiveTranslationMessage(dst_bundle, msg.key, msg.value, t_history, now, msg.taken_from_default, msg.position, msg.category, msg.from_developer)
             db.session.add(active_t)
-            db.session.commit()
+            try:
+                db.session.commit()
+            except:
+                db.session.rollback()
+                raise
         elif existing_translation.taken_from_default and not msg.taken_from_default:
             # Merge it
             t_history = TranslationMessageHistory(dst_bundle, msg.key, msg.value, msg.history.user, now, existing_translation.history.id, msg.taken_from_default)
@@ -417,7 +444,11 @@ def _merge_bundle(src_bundle, dst_bundle):
             active_t = ActiveTranslationMessage(dst_bundle, msg.key, msg.value, t_history, now, msg.taken_from_default, msg.position, msg.category, msg.from_developer)
             db.session.add(active_t)
             db.session.delete(existing_translation)
-            db.session.commit()
+            try:
+                db.session.commit()
+            except:
+                db.session.rollback()
+                raise
 
 def _deep_copy_translations(old_translation_url, new_translation_url):
     """Given an old translation of a URL, take the old bundles and copy them to the new one."""
@@ -438,7 +469,11 @@ def start_synchronization():
     now = datetime.datetime.utcnow()
     sync_log = TranslationSyncLog(now, None)
     db.session.add(sync_log)
-    db.session.commit()
+    try:
+        db.session.commit()
+    except:
+        db.session.rollback()
+        raise
     db.session.refresh(sync_log)
     return sync_log.id
 
@@ -447,7 +482,11 @@ def end_synchronization(sync_id):
     sync_log = db.session.query(TranslationSyncLog).filter_by(id = sync_id).first()
     if sync_log is not None:
         sync_log.end_datetime = now
-        db.session.commit()
+        try:
+            db.session.commit()
+        except:
+            db.session.rollback()
+            raise
 
 def get_latest_synchronizations():
     latest_syncs = db.session.query(TranslationSyncLog)[-10:]
@@ -482,8 +521,12 @@ def update_user_status(language, target, app_url, user):
         db.session.add(active_user)
     else:
         active_user.update_last_check()
-
-    db.session.commit()
+    
+    try:
+        db.session.commit()
+    except:
+        db.session.rollback()
+        raise
 
 def get_user_status(language, target, app_url, user):
     FORMAT = "%Y-%m-%dT%H:%M:%SZ"
